refactor: move device info collection into library layer (#1240)
